requirements.txt
streamlit
ultralytics
pillow
pandas
numpy
scikit-learn